+++ clients/admin-ui/src/features/common/nav/v2/routes.ts
@@ -20,6 +20,7 @@ export const DATASET_COLLECTION_SUBFIELD_DETAIL_ROUTE =
   "/dataset/[datasetId]/[collectionName]/[...subfieldNames]";
 
 // Detection and discovery
+export const ACTION_CENTER_ROUTE = "/data-discovery/action-center";
 export const DETECTION_DISCOVERY_ACTIVITY_ROUTE = "/data-discovery/activity";
 export const DATA_DETECTION_ROUTE = "/data-discovery/detection";
 export const DATA_DETECTION_ROUTE_DETAIL =